refactor(fields): update imports to use FilterFormFieldModel
