+++ packages/core/client/src/flow/models/filter-blocks/form/fields/RadioGroupFilterFormEditableFieldModel.tsx
@@ -8,7 +8,7 @@
  */
 
 import { Select } from '@formily/antd-v5';
-import { FilterFormFieldModel } from './FilterFormEditableFieldModel';
+import { FilterFormFieldModel } from './FilterFormFieldModel';
 
 export class RadioGroupFilterFormEditableFieldModel extends FilterFormFieldModel {
   static readonly supportedFieldInterfaces = ['radioGroup'];
