+++ packages/core/client/src/flow/models/filter-blocks/form/fields/SelectFilterFormEditableFieldModel.tsx
@@ -8,7 +8,7 @@
  */
 
 import { Select } from '@formily/antd-v5';
-import { FilterFormFieldModel } from './FilterFormEditableFieldModel';
+import { FilterFormFieldModel } from './FilterFormFieldModel';
 
 export class SelectFilterFormEditableFieldModel extends FilterFormFieldModel {
   static readonly supportedFieldInterfaces = ['select', 'multipleSelect'];
