+++ packages/core/client/src/flow/models/filter-blocks/form/fields/TimeFilterFormEditableFieldModel.tsx
@@ -8,7 +8,7 @@
  */
 
 import { TimePicker } from '@formily/antd-v5';
-import { FilterFormFieldModel } from './FilterFormEditableFieldModel';
+import { FilterFormFieldModel } from './FilterFormFieldModel';
 
 export class TimeFilterFormEditableFieldModel extends FilterFormFieldModel {
   static readonly supportedFieldInterfaces = ['time'];
