+++ packages/core/client/src/flow/models/filter-blocks/form/fields/index.ts
@@ -8,7 +8,7 @@
  */
 
 export * from './InputFilterFormEditableFieldModel';
-export * from './FilterFormEditableFieldModel';
+export * from './FilterFormFieldModel';
 export * from './CheckboxFilterFormEditableFieldModel';
 export * from './CheckboxGroupFilterFormEditableFieldModel';
 export * from './SelectFilterFormEditableFieldModel';
